refactor:작품 추가,수정 모달 공통 컴포넌트로 추출
